fix: don't adjust mediaPresentationDuration by timescale for segment duration when using SegmentBase (#94)
When using SegmentBase, there is only one segment. If the duration is
not present, sourceDuration is used, which comes from
mediaPresentationDuration.

Previously, sourceDuration was being divided by the timescale, but
mediaPresentationDuration is defined in the DASH spec as s:duration,
which follows ISO 8601. It doesn't need to be adjusted based on the
timescale.

This change fixes the use of sourceDuration for SegmentBase by not
performing that division.