feat: support multi-period live DASH with changing periods
In order to handle live DASH manifests where periods are changed, added,
and removed, merging is used to ensure that the new manifest is a
continuation of the last parsed manifest. In addition, playlists that
need to be excluded are identified in the returned object.

BREAKING CHANGE: the top level API of this module no longer returns the
manifest alone