feat: support multi-period live DASH with changing periods
In order to handle live DASH manifests where periods are changed, added, and removed,
merging is used to ensure that the new manifest is a continuation of the last parsed
manifest.

In order to handle live cases, the last MPD must be provided via the `lastMpd` option.

BREAKING CHANGE: period start time is now used as the timeline instead of period index