feat: support multi-period live DASH with changing periods (#152)
In order to handle live DASH manifests where periods are changed, added, and removed,
merging is used to ensure that the new manifest is a continuation of the last parsed
manifest. To achieve this, the previously loaded manifest should be provided via the new `previousManifest` option.


Multiperiod SIDX (in both live and VOD) and live SIDX where the SIDX
properties change on refresh, are not yet supported in mpd-parser. To
prevent exceptions, return early from merging logic on SIDX detection,
and add a TODO comment for the future.

BREAKING CHANGE: period start time (periodStart) is now used as the timeline instead of period index (periodIndex)